Add logic to figure out the timezone of the user
This can be used to figure out where in the world the user resides. We
need to take special care here because some platforms don't have a
function that we require. The timezone isn't critical data so if we fail
to determine it we simply ignore it using a try/catch block.