Implement pageview tracking
This commit sets up the actual tracking of pageviews. It implements this
both for the case where the page is reloaded on each request
(traditional navigation) as well as when the history API is used for
single-page navigation.

If possible it submits all data at once using the beacon API. If not
possible it just posts the data directly.