Set up docker webserver
This sets up a docker webserver that can serve the static JavaScript
code and forward all other requests to the Node server. This mimicks how
this set-up would run in production. It has the added benefit of
allowing us to use nginxproxy to set a hostname.

The predev command is adjusted to build the project so that `webpack
--watch` only begins watching. This also ensures that the build output
is already present when nodemon is started. If this is not done then
nodemon will throw an error because webpack is not done with bundling.