Update HapiJS and Joi
The leadership for HapiJS has changed and Joi has been spun out as a
result to remain under the original maintainer.

The new leadership seems to be fine so we should be safe to continue
using Hapi without a costly migration.

More info can be read in the following issues:

- https://github.com/sideway/joi/issues/2411
- https://github.com/hapijs/hapi/issues/4111
- https://github.com/hapijs/hapi/issues/4113

@hapi/hapi 20.0.0 is a drop-in replacement for 19.x that fixes a
security issue. The major version change is to reflect the new
management.

Release notes: https://github.com/hapijs/hapi/issues/4113